Support for JSON as string
A JSONValue object can now be created directly from a JSON string,
without having to go through an NSData object.

This is done by providing the external parameter name jsonString to the
initializer - consistency could be improved by having named parameters
for the NSData initializer,  as well, but that would be a breaking change.