The ‘isBool’ function on the NSNumber extension uses ‘trueObjCType’, which represents the type encoding for ‘true’ represented as an NSNumber, and ‘falseObjCType’, which represents the type encoding for ‘false’ represented as an NSNumber. Both type encodings are equal to ‘c’. However, an NSNumber that wraps around a value of type Int8 also has a type encoding of ‘c’. Therefore, a user who passes in a value of 0 or 1 (of type Int8) into JSON will expect to see the numbers 0 or 1 as the output because that was the intent. Unfortunately, that value will be wrapped into NSNumber and be considered a bool, so the user will see ‘true’ or ‘false’ instead.
Therefore, when setting a new value of type Int8 on the computed properties ‘int8’ or ‘int8Value’, the new value should be used to directly set ‘object’, which should then check for the Int8 type and use an Int instead. This is fine because an Int8 converted to an Int will always work and retain the same behavior of JSON, and also because the user’s intent is to use a number, not a boolean.

Having said that, the tests for Int8 don’t test the right thing. The user’s intent is to use an integer, which is a number. However, the test assumes hardware implementation and accepts the fact that it can either be true or a number, which is unintuitive. Int8 is a pragmatic refinement to the semantic meaning of an integer, and should therefore be used as a number like the user intends, not a boolean.