Fix init(data:) error handling:
The problem is that when users call `let json = JSON(data: data)`, they are not informed of the error, relying on the presence of `NSNull` value. That sort of use of non-standard sentinel value is discouraged. And the reliance of the `NSErrorPointer` is a pattern that pre-dates Swift 2 (and was intended for interacting with Objective-C API, not to be adopted within a Swift API itself). Bottom line, the error handling in  `init(data:options:error:)` was distinctly unswifty. This attempts to remedy that.

1. Change init(data:options:) to throw errors, removing the `NSErrorPointer` reference.

2. Changed existing code that called `init(data:)` to reflect this error throwing behavior.

3. Also changed default reading options to be `[]` rather than `.allowFragments`, so that it's more consistent with standard `JSONSerialization` behavior. Also, I want to stop the perpetuation of the use of `.allowFragments` as a matter of standard practice, because generally if you have a fragment of JSON, you really want to know about that error.

4. Added test to demonstrate the catching of thrown error.